Fix after code review
