+++ core/src/luigi-config/luigi-config.js
@@ -2,11 +2,11 @@
 import {
   saveCurrentLocation,
   getPreviousLocation,
-  setLimitExceededErrorsMessages
 } from './navigation/navigation-helpers';
 import { communication } from './communication';
 import { config } from './config';
 import { navigation, getNavigationData, getToken } from './navigation/navigation-data-init';
+import './luigi-listeners';
 
 
 function getFreshKeys() {
@@ -14,22 +14,6 @@ function getFreshKeys() {
   // until dex has possibility of no-cache
   return fetch('https://dex.' + config.domain + '/keys', { cache: 'no-cache' });
 }
-
-
-window.addEventListener('message', e => {
-  const SHOW_SYSTEM_NAMESPACES_CHANGE_EVENT =
-    'showSystemNamespacesChangedEvent';
-
-  if (e.data && e.data.msg === 'luigi.refresh-context-switcher') {
-    window.Luigi.cachedNamespaces = null;
-  } else if (e.data && e.data.msg === SHOW_SYSTEM_NAMESPACES_CHANGE_EVENT) {
-    Luigi.customMessages().sendToAll({
-      id: SHOW_SYSTEM_NAMESPACES_CHANGE_EVENT,
-      showSystemNamespaces: e.data.showSystemNamespaces
-    });
-  }
-});
-
 
 getFreshKeys()
 
@@ -110,50 +94,5 @@ const luigiConfig = {
   }
 };
 Luigi.setConfig(luigiConfig);
-
-window.addEventListener('message', e => {
-  if (e.data.msg && e.data.msg === 'console.quotaexceeded') {
-    const namespace = e.data.namespace;
-    const data = e.data.data;
-    let limitHasBeenExceeded;
-    let limitExceededErrors;
-    if (data && data.resourceQuotasStatus) {
-      limitHasBeenExceeded = data.resourceQuotasStatus.exceeded;
-    }
-    if (
-      data &&
-      data.resourceQuotasStatus &&
-      data.resourceQuotasStatus.exceededQuotas &&
-      data.resourceQuotasStatus.exceededQuotas.length > 0
-    ) {
-      limitExceededErrors = setLimitExceededErrorsMessages(
-        data.resourceQuotasStatus.exceededQuotas
-      );
-      const linkdata = {
-        goToResourcesConfig: {
-          text: 'Resources Configuration',
-          url: `/home/namespaces/${namespace}/resources`
-        }
-      };
-      let errorText = `Error ! The following resource quota limit has been exceeded by the given resource:<br>`;
-      limitExceededErrors.forEach(error => {
-        errorText += `-${error}<br>`;
-      });
-      errorText += `See {goToResourcesConfig} for details.`;
-      const settings = {
-        text: errorText,
-        type: 'error',
-        links: linkdata
-      };
-      window.postMessage(
-        {
-          msg: 'luigi.ux.alert.show',
-          data: { settings }
-        },
-        '*'
-      );
-    }
-  }
-});
